+++ js/main.js
@@ -1,0 +1,3 @@
+import {createData} from './create-data.js';
+
+window.console.log(createData());
