+++ js/util.js
@@ -1,22 +1,5 @@
-//Случайное целое число в диапазоне, включая минимальное и максимальное.
-function getRandomInRange(min, max) {
-  let errorMessage = '';
-
-  if (min < 0) {
-    errorMessage = 'Минимальное значение меньше нуля';
-  } else if (max <= min) {
-    errorMessage = 'Минимальное значение больше или равно максимальному';
-  }
-
-  if (!errorMessage) {
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-  }
-
-  window.console.error(errorMessage);
-}
-
 //Функция для проверки максимальной длины строки
-function checkLengthStr(str, maxLengthForStr) {
+const checkLengthStr = (str, maxLengthForStr) => {
   let result = true;
 
   if (maxLengthForStr < str.length) {
@@ -24,6 +7,8 @@ function checkLengthStr(str, maxLengthForStr) {
   }
 
   return result;
-}
+};
 
 checkLengthStr('123', 2);
+
+export {checkLengthStr};
